requirements.txt
environs==11.2.1
python-telegram-bot==21.10
requests==2.32.3
